requirements.txt
Django==1.6.5
dj-database-url==0.3.0
dj-static==0.0.6
django-extensions==1.3.11
django-filter==0.7
django-toolbelt==0.0.1
djangorestframework==2.4.2
gunicorn==19.1.1
psycopg2==2.5.4
six==1.7.3
static3==0.5.1
wsgiref==0.1.2
